Made older version of QUTMS_Windows successfully compile.
The function returnPortNumber() now checks that the length of the portList[] array is greater than zero before returning it's length minus one.
If the portList array is less than zero, the function returns null which causes the program to return a dialog indicating a port error but allows the program to run nonetheless when the program would previously refuse to run.