Implemented artificial packet response.
The status of the sixth channel in the Power Management screen now changes depending on the number contained in the message byte of a test packet.
This shows the GUI and the Packet class are working so further work can be done as more information on what the power management packets actually contain become avaliable.